+++ lib/src/i_set.dart
@@ -1,21 +1,22 @@
 part of 'i_collection.dart';
 
 /// Immutable Set.
-class ISet<E> extends _ICollection<Set<E>>
+class ISet<E> extends _ICollection<HashSet<E>>
     with
-        _ImmutableIterableMixin<E, Set<E>>,
-        _ImmutableSetMixin<E>,
-        _MutableWithThrowSetMixin<E> {
+        _ImmutableIterableMixin<E, HashSet<E>>,
+        _ImmutableHashSetMixin<E>,
+        _MutableWithThrowHashSetMixin<E> {
   /// Creates an ISet instance from [source].
   /// Use only if you are sure that the [source] will not be modified in the future.
   @literal
-  const ISet(Set<E> source) : super(source);
+  const ISet(HashSet<E> source) : super(source);
 
   /// Creates an ISet instance from [source] by copying it.
-  ISet.of(Iterable<E> source) : super(Set.of(source));
+  ISet.of(Iterable<E> source) : super(HashSet.of(source));
 }
 
-mixin _ImmutableSetMixin<E> on _ICollection<Set<E>> implements Set<E> {
+mixin _ImmutableHashSetMixin<E> on _ICollection<HashSet<E>>
+    implements HashSet<E> {
   @override
   Set<R> cast<R>() => _source.cast();
 
@@ -35,7 +36,7 @@ mixin _ImmutableSetMixin<E> on _ICollection<Set<E>> implements Set<E> {
   Set<E> union(Set<E> other) => _source.union(other);
 }
 
-mixin _MutableWithThrowSetMixin<E> implements Set<E> {
+mixin _MutableWithThrowHashSetMixin<E> implements HashSet<E> {
   @alwaysThrows
   R _throw<R>() {
     throw UnsupportedError('ISet<$E> cannot be modified.');